Cleanups

+++ tests/src/test_obj0.zig
@@ -1,12 +1,12 @@
 const std = @import("std");
 const beacon = @import("bofapi").beacon;
 
-fn func(msg: []const u8) i32 {
+fn func(msg: []const u8) u8 {
     std.debug.print("debug: test {s}\n", .{msg});
     return 0;
 }
 
-pub export fn go(arg_data: ?[*]u8, arg_len: i32) callconv(.C) i32 {
+pub export fn go(arg_data: ?[*]u8, arg_len: i32) callconv(.C) u8 {
     _ = arg_data;
     _ = arg_len;
     _ = beacon.printf(0, "--- test_obj0.zig ---\n");
