+++ tests/src/test_obj4.zig
@@ -2,7 +2,7 @@ const std = @import("std");
 const fmt = std.fmt;
 const beacon = @import("bofapi").beacon;
 
-pub export fn go(arg_data: ?[*]u8, arg_len: i32) callconv(.C) i32 {
+pub export fn go(arg_data: ?[*]u8, arg_len: i32) callconv(.C) u8 {
     var parser = beacon.datap{};
 
     _ = beacon.printf(0, "--- test_obj4.zig ---\n");
